pyproject.toml
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = ["mypy>=1.8.0,<2.0", "flake8>=7.0.0,<8.0", "black>=24.1.1,<25.0", "pytest>=8.0.0,<9.0"]

[project]
name = "voicebrief"
version = "0.1.0"
description = "Converts video / audio conversations to text and subsequently provides a summary into a managable report"
license = "MIT"
readme = "README.md"
requires-python = ">=3.11,<4.0"
dependencies = [
  "openai>=1.7.2,<2.0",
  "moviepy>=1.0.3,<2.0",
  "pydub>=0.25.1,<1.0",
  "python-dotenv>=1.0.0,<2.0",
  "tiktoken>=0.5.2,<1.0",
  # Python 3.13 removed stdlib 'audioop'. pydub can use this LTS backport.
  "audioop-lts; python_version >= '3.13'",
]

[[project.authors]]
name = "Iwan van der Kleijn"
email = "iwanvanderkleijn@gmail.com"

[project.scripts]
check-all = "devenv.run_checks:run_checks"

[project.optional-dependencies]
dev = ["mypy>=1.8.0,<2.0", "flake8>=7.0.0,<8.0", "black>=24.1.1,<25.0", "pytest>=8.0.0,<9.0"]
